Added a helper function to call multiple control actions at once in the webview frontend and implemented for the settings menu.
Removed the ability to remove a profile if it is the currently selected one or if it is the only one in both frontends.

Backend:
Disallowed the removal of the current active profile in the backend.
Fixed the active_profile index being incorrect after removing a profile sometimes.
Fixed program crash after removing profile at any of the ends.